ChatbotMVP_01/backend/requirements.latest.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-core
langchain-openai
langchain-community
langchain-aws
boto3
